requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
python-docx==1.2.0
PyMuPDF==1.26.3
pdfplumber==0.11.7
pytesseract==0.3.13
google-cloud-vision==3.13.0
Pillow==11.3.0
opencv-python-headless==4.12.0.88
numpy==2.2.6
spacy==3.8.7
transformers==4.55.2
torch==2.8.0
celery==5.5.3
redis==6.4.0
httpx==0.28.1
pytest==8.4.1